+++ apps/vscode-extension/src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { clearVariablesState } from './utils';
 import {
+  AnalyticsCommands,
   DecoratorService,
   DemoApi,
   DemoCreator,
@@ -13,6 +14,7 @@ import {
   NotesService,
   PdfExportService,
   Slides,
+  SponsorService,
   UriHandler,
   TextTypingService,
   TerminalService,
@@ -61,6 +63,8 @@ export async function activate(context: vscode.ExtensionContext) {
   ImportService.register();
   TerminalService.register();
   InputService.registerCommands();
+  AnalyticsCommands.registerCommands();
+  SponsorService.init(context);
 
   console.log(`${Config.title} is active!`);
 }