+++ apps/vscode-extension/src/webview/WebviewHtml.ts
@@ -5,8 +5,8 @@ export const WebviewHtml = `<!doctype html>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Demo Time Webview</title>
-    <script type="module" crossorigin src="./assets/index-BgTuUnMW.js"></script>
-    <link rel="stylesheet" crossorigin href="./assets/index-CQi4U1yH.css">
+    <script type="module" crossorigin src="./assets/index-B-7r10LP.js"></script>
+    <link rel="stylesheet" crossorigin href="./assets/index-DhkUc8AL.css">
   </head>
   <body>
     <div id="root"></div>