+++ packages/common/src/constants/Config.ts
@@ -10,6 +10,7 @@ export const Config = {
   presentationMode: {
     previousEnabled: 'previousEnabled',
   },
+  zoom: 'zoomLevel',
   highlight: {
     borderColor: 'highlightBorderColor',
     background: 'highlightBackground',
@@ -52,5 +53,8 @@ export const Config = {
   remote: {
     showScreenshot: 'remote.showScreenshot',
     showNotes: 'remote.showNotes',
+  },
+  analytics: {
+    enabled: 'analytics.enabled',
   },
 };