+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## [1.13.0] - unreleased
+
+- [#339](https://github.com/estruyf/vscode-demo-time/issues/339): Prevent
+  unnecessary zoom when consecutive highlights are performed
+- [#340](https://github.com/estruyf/vscode-demo-time/issues/340): Fix an issue
+  with the `path` label for the `executeScript` action in the step editor
+- [#341](https://github.com/estruyf/vscode-demo-time/issues/341): Added the
+  `args` property to the `executeScript` action to allow passing command-line
+  arguments to the script being executed
+- [#343](https://github.com/estruyf/vscode-demo-time/issues/343): added the zoom
+  actions: `zoomIn`, `zoomOut`, and `zoomReset` to control the editor zoom level
+  during your demo
+- [#348](https://github.com/estruyf/vscode-demo-time/issues/348): Added the
+  analytics PRO feature
+
 ## [1.12.0] - 2025-12-16
 
 - **BREAKING**: Changed project license from MIT to Apache License 2.0. See